iteration1-create-ticket_v1
Created the supabase edge function based on the current database schematic. The function purpose is to create a ticket and assign it to the staff with the least amount of "pending" tickets, it returns nothing and will console the result. Please remember to edit the .env for own personal usage.